FA-40: Cert generation and changes for TLS transport, including fixing a DER vs PEM issue.
